ATR {TTR}	R Documentation

True Range / Average True Range

Description

True range (TR) is a measure of volatility of a High-Low-Close series; average true range (ATR) is a Welles Wilder's style moving average of the TR. Developed by J. Welles Wilder in 1978.

Usage

ATR(HLC, n = 14, maType, ...)

Arguments

HLC
Object that is coercible to xts or matrix and contains High-Low-Close prices.

n
Number of periods for moving average.

maType
A function or a string naming the function to be called.

...
Other arguments to be passed to the maType function.

Details

TR incorporates yesterday's close in the calculation (high minus low). E.g. if yesterday's close was higher than today's high, then the TR would equal yesterday's close minus today's low.

The ATR is a component of the Welles Wilder Directional Movement Index (DX, ADX).

Value

A object of the same class as HLC or a matrix (if try.xts fails) containing the columns:

tr

The true range of the series.

atr

The average (as specified by ma) true range of the series.

trueHigh

The true high of the series.

trueLow

The true low of the series.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.fmlabs.com/reference/TR.htm
http://www.fmlabs.com/reference/ATR.htm
https://www.metastock.com/Customer/Resources/TAAZ/?p=35
https://www.linnsoft.com/techind/true-range-tr
http://www.stockcharts.com/school/doku.php?id=chart_school:technical_indicators:average_true_range_atr

See Also

See EMA, SMA, etc. for moving average options; and note Warning section. See DX, which uses true range. See chaikinVolatility for another volatility measure.

Examples

data(ttrc)
atr <- ATR(ttrc[,c("High","Low","Close")], n=14)

[Package TTR version 0.23-2 Index]