CLV {TTR}	R Documentation

Close Location Value

Description

The Close Location Value (CLV) relates the day's close to its trading range.

Usage

CLV(HLC)

Arguments

HLC
Object that is coercible to xts or matrix and contains High-Low-Close prices.

Details

The CLV will fall in a range of -1 to +1. If the CLV is +/-1, the close is at the high/low; if the CLV is 0, the close is directly between the high and low.

Value

A object of the same class as HLC or a vector (if try.xts fails) containing the Close Location Values of a High-Low-Close price series.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:accumulation_distribution_line

See Also

See chaikinAD, which uses CLV.

Examples

data(ttrc)
clv <- CLV(ttrc[,c("High","Low","Close")])

[Package TTR version 0.23-2 Index]