CMF {TTR}	R Documentation

Chaikin Money Flow

Description

Chaikin Money Flow compares total volume over the last n time periods to total volume times the Close Location Value (CLV) over the last n time periods. Developed by Marc Chaikin.

Usage

CMF(HLC, volume, n = 20)

Arguments

HLC
Object that is coercible to xts or matrix and contains High-Low-Close prices.

volume
Vector or matrix of volume observations corresponding to the HLC object.

n
Number of periods to use.

Details

Chaikin Money Flow is calculated by taking dividing the sum of the Chaikin Accumulation / Distribution line over the past n periods by the sum of volume over the past n periods.

Value

A object of the same class as HLC and volume or a vector (if try.xts fails) containing the Chaikin Money Flow values.

Note

When Chaikin Money Flow is above/below +/- 0.25 it is a bullish/bearish signal. If Chaikin Money Flow remains below zero while the price is rising, it indicates a probable reversal.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.fmlabs.com/reference/ChaikinMoneyFlow.htm
https://www.linnsoft.com/techind/chaikin-money-flow-cmf
http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:chaikin_money_flow_cmf

See Also

See CLV, and chaikinAD.

Examples

data(ttrc)
cmf <- CMF(ttrc[,c("High","Low","Close")], ttrc[,"Volume"])

[Package TTR version 0.23-2 Index]