CMO {TTR}	R Documentation

Chande Momentum Oscillator

Description

The Chande Momentum Oscillator (CMO) is a modified RSI. Developed by Tushar S. Chande.

Usage

CMO(x, n = 14)

Arguments

x
Price, volume, etc. series that is coercible to xts or matrix.

n
Number of periods to use.

Details

The CMO divides the total movement by the net movement ([up - down] / [up + down]), where RSI divides the upward movement by the net movement (up / [up + down]).

Value

A object of the same class as x or a vector (if try.xts fails) containing Chande Momentum Oscillator values.

Note

There are several ways to interpret the CMO:

Values over/under +/- 50 indicate overbought/oversold conditions.

High CMO values indicate strong trends.

When the CMO crosses above/below a moving average of the CMO, it is a buy/sell signal.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.fmlabs.com/reference/CMO.htm

See Also

See RSI.

Examples

data(ttrc)
cmo <- CMO(ttrc[,"Close"])

[Package TTR version 0.23-2 Index]