DVI {TTR}	R Documentation

DV Intermediate Oscillator

Description

The DV Intermediate oscillator (DVI) is a very smooth momentum oscillator that can also be used as a trend indicator. Created by David Varadi.

Usage

DVI(price, n = 252, wts = c(0.8, 0.2), smooth = 3, magnitude = c(5, 100,
  5), stretch = c(10, 100, 2), exact.multiplier = 1)

Arguments

price
Price series that is coercible to xts or matrix.

n
Number of periods for the percent rank.

wts
The weight given to the smoothed returns (magnitude) component and the up/down days (stretch) component, respectively.

smooth
The number of periods to smooth price.

magnitude
A set of 3 periods used to smooth magnitude.

stretch
A set of 3 periods used to smooth stretch.

exact.multiplier
The weight applied to identical values in the window. See runPercentRank.

Details

The DVI combines smoothed returns over different time windows and the relative number of up versus down days (stretch) over different time windows.

Value

A object of the same class as price or a vector (if try.xts fails) containing the DVI values.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://cssanalytics.wordpress.com/2009/12/13/what-is-the-dvi/
http://marketsci.wordpress.com/2010/07/27/css-analytics%E2%80%99-dvi-indicator-revealed/

Examples

data(ttrc)
dvi <- DVI(ttrc[,"Close"])

[Package TTR version 0.23-2 Index]