EMV {TTR}	R Documentation

Arms' Ease of Movement Value

Description

Arms' Ease of Movement Value (EMV) emphasizes days where the security moves easily and minimizes days where the security does not move easily. Developed by Richard W. Arms, Jr.

Usage

EMV(HL, volume, n = 9, maType, vol.divisor = 10000, ...)

Arguments

HL
Object that is coercible to xts or matrix and contains High-Low prices.

volume
Vector or matrix of volume observations corresponding to the HL object.

n
Number of periods for moving average.

maType
A function or a string naming the function to be called.

vol.divisor
An increment to make the results larger and easier to work with.

...
Other arguments to be passed to the maType function.

Details

The EMV is calculated by dividing the midpoint ([high + low]/2) move by the 'Box Ratio' (volume divided by the high minus low).

Value

A object of the same class as HL and volume or a matrix (if try.xts fails) containing the columns:

emv

The ease of movement values.

emvMA

The smoothed (as specified by ma) ease of movement values.

Note

A buy/sell signal is generated when the EMV crosses above/below zero. When the EMV hovers around zero, there are small price movements and/or high volume, and the price is not moving easily.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.fmlabs.com/reference/ArmsEMV.htm
https://www.metastock.com/Customer/Resources/TAAZ/?p=51
https://www.linnsoft.com/techind/arms-ease-movement

See Also

See EMA, SMA, etc. for moving average options; and note Warning section.

Examples

data(ttrc)
emv <- EMV(ttrc[,c("High","Low")], ttrc[,"Volume"])

[Package TTR version 0.23-2 Index]