GMMA {TTR}	R Documentation

Guppy Multiple Moving Averages

Description

Calculate the Guppy Multiple Moving Average of a series.

Usage

GMMA(x, short = c(3, 5, 8, 10, 12, 15), long = c(30, 35, 40, 45, 50, 60),
  maType)

Arguments

x
Price, volume, etc. series that is coercible to xts or matrix.

short
Vector of short-term periods.

long
Vector of long-term periods.

maType
Either:

A function or a string naming the function to be called.

A list with the first component like (1) above, and additional parameters specified as named components. See Examples.

Details

The Guppy Multiple Moving Average signals a changing trend when the short and long groups of moving averages intersect. An up/down trend exists when the short/long-term moving averages are greater than the long/short-term averages.

Value

A object of the same class as x or price or a vector (if try.xts fails) containing the Guppy Multiple Moving Average.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.investopedia.com/terms/g/guppy-multiple-moving-average.asp

See Also

See aroon, CCI, ADX, VHF, TDI for other indicators that measure trend direction/strength.

Examples

data(ttrc)
gmma <- GMMA(ttrc[,"Close"])

[Package TTR version 0.23-2 Index]