OBV {TTR}	R Documentation

On Balance Volume (OBV)

Description

On Balance Volume (OBV) is a measure of the money flowing into or out of a security. It is similar to Chaikin Accumulation / Distribution.

Usage

OBV(price, volume)

Arguments

price
Price series that is coercible to xts or matrix.

volume
Volume series that is coercible to xts or matrix, that corresponds to price object.

Details

OBV is calculated by adding (subtracting) each day's volume to a running cumulative total when the security's price closes higher (lower).

Value

A object of the same class as price and volume or a vector (if try.xts fails) containing the OBV values.

Note

OBV is usually compared with the price chart of the underlying security to look for divergences/confirmation.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.fmlabs.com/reference/OBV.htm
https://www.metastock.com/Customer/Resources/TAAZ/?p=82
https://www.linnsoft.com/techind/balance-open-interest
http://www.stockcharts.com/school/doku.php?id=chart_school:technical_indicators:on_balance_volume_obv

See Also

See chaikinAD.

Examples

data(ttrc)
obv <- OBV(ttrc[,"Close"], ttrc[,"Volume"])

[Package TTR version 0.23-2 Index]