SNR {TTR}	R Documentation

Signal to Noise Ratio

Description

The n-day SNR for a given market is calculated by taking the absolute price change over an n-day period and dividing it by the average n-day volatility.

Usage

SNR(HLC, n, ...)

Arguments

HLC
Object that is coercible to xts or matrix and contains High-Low-Close prices.

n
Number of periods for moving average.

...
Other arguments to be passed to ATR.

Details

SNR = abs(Cl - lag(Cl,n)) / ATR(HLC, n)$atr

Using average true range as the volatility measure captures more of the intraday and overnight volatility in a way that a measurement of Close-to-Close price change does not.

The interpretation is then relatively intuitive: an SNR value of five indicates that the market has moved five times the volatility (average true range) over the given look-back period.

Value

A object of the same class as HLC or a matrix (if try.xts fails) containing the signal to noise ratio.

Author(s)

Peter Carl

References

Skeggs, James and Hill, Alex (2015). Back in Black Part 2: The Opportunity Set for Trend Following.

[Package TTR version 0.23-2 Index]