TDI {TTR}	R Documentation

Trend Detection Index

Description

The Trend Detection Index (TDI) attempts to identify starting and ending trends. Developed by M. H. Pee.

Usage

TDI(price, n = 20, multiple = 2)

Arguments

price
Price series that is coercible to xts or matrix.

n
Number of periods to use.

multiple
Multiple used to calculate (2).

Details

The TDI is the (1) absolute value of the n-day sum of the n-day momentum, minus the quantity of (2) multiple*n-day sum of the absolute value of the n-day momentum, minus (3) n-day sum of the absolute value of the n-day momentum.

I.e. TDI = (1) - [ (2) - (3) ]

The direction indicator is the sum of the n-day momentum over the last n days.

See URL in references section for further details.

Value

A object of the same class as price or a matrix (if try.xts fails) containing the columns:

tdi

The Trend Detection Index.

di

The Direction Indicator.

Note

Positive/negative TDI values signal a trend/consolidation. A positive/ negative direction indicator signals a up/down trend. I.e. buy if the TDI and the direction indicator are positive, and sell if the TDI is positive while the direction indicator is negative.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
https://www.linnsoft.com/techind/trend-detection-index-tdi

See Also

See aroon, CCI, ADX, VHF, GMMA for other indicators that measure trend direction/strength.

Examples

data(ttrc)
tdi <- TDI(ttrc[,"Close"], n=30)

[Package TTR version 0.23-2 Index]