adjRatios {TTR}	R Documentation

Split and dividend adjustment ratios

Description

Create split and dividend adjustment ratio vectors.

Usage

adjRatios(splits, dividends, close)

Arguments

splits
Split series that is coercible to xts.

dividends
Dividend series that is coercible to xts.

close
Close price series that is coercible to xts.

Details

If only splits is provided, the resulting object will only have as many observations as splits.

If splits and close are provided, the resulting object will have as many observations as max(NROW(splits), NROW(close)).

close is required if dividends is provided.

Value

A xts object containing the columns:

Split

The split adjustment ratio.

Div

The dividend adjustment ratio.

Author(s)

Joshua Ulrich

[Package TTR version 0.23-2 Index]