aroon {TTR}	R Documentation

Aroon

Description

The Aroon indicator attempts to identify starting trends. The indicator consists of up and down lines, which measure how long it has been since the highest high/lowest low has occurred in the last n periods. Developed by Tushar Chande in 1995.

Usage

aroon(HL, n = 20)

Arguments

HL
Object that is coercible to xts or matrix and contains either a High-Low price series, or a Close price series.

n
Number of periods to use in the calculation.

Details

Aroon up (down) is the elapsed time, expressed as a percentage, between today and the highest (lowest) price in the last n periods. If today's price is a new high (low) Aroon up (down) will be 100. Each subsequent period without another new high (low) causes Aroon up (down) to decrease by (1 / n) x 100.

Value

A object of the same class as HL or a matrix (if try.xts fails) containing the columns:

aroonUp

The Aroon up indicator.

aroonDn

The Aroon down indicator.

oscillator

The Aroon oscillator (aroonUp - aroonDn).

Note

If High-Low prices are given, the function calculates the max/min using the high/low prices. Otherwise the function calculates the max/min of the single series.

Up (down) trends are indicated when the aroonUp(Dn) is between 70 and 100. Strong trends are indicated when when the aroonUp(Dn) is above 70 while the aroonDn(Up) is below 30. Also, crossovers may be useful.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.fmlabs.com/reference/Aroon.htm
http://www.fmlabs.com/reference/AroonOscillator.htm
https://www.linnsoft.com/techind/aroon-arn
http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:aroon

See Also

See CCI, ADX, TDI, VHF, GMMA for other indicators that measure trend direction/strength.

Examples

## Get Data and Indicator ##
data(ttrc)
trend <- aroon( ttrc[,c("High", "Low")], n=20 )

[Package TTR version 0.23-2 Index]