chaikinAD {TTR}	R Documentation

Chaikin Accumulation / Distribution

Description

The Chaikin Accumulation / Distribution (AD) line is a measure of the money flowing into or out of a security. It is similar to On Balance Volume (OBV). Developed by Marc Chaikin.

Usage

chaikinAD(HLC, volume)

Arguments

HLC
Object that is coercible to xts or matrix and contains High-Low-Close prices.

volume
Vector or matrix of volume observations corresponding to the HLC object.

Details

The AD line is similar to OBV; the difference is that OBV sums volume multiplied by +/- 1 if the close is higher/lower than the previous close, while the AD line multiplies volume by the close location value (CLV).

Value

A object of the same class as HLC and volume or a vector (if try.xts fails) containing the accumulation / distribution values.

Note

The Accumulation/Distribution Line is interpreted by looking for a divergence in the direction of the indicator relative to price.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.fmlabs.com/reference/AccumDist.htm
https://www.metastock.com/Customer/Resources/TAAZ/?p=27
https://www.linnsoft.com/techind/accumulation-distribution
http://www.stockcharts.com/school/doku.php?id=chart_school:technical_indicators:accumulation_distribution_line

See Also

See OBV, and CLV.

Examples

data(ttrc)
ad <- chaikinAD(ttrc[,c("High","Low","Close")], ttrc[,"Volume"])

[Package TTR version 0.23-2 Index]