chaikinVolatility {TTR}	R Documentation

Chaikin Volatility

Description

Chaikin Volatility measures the rate of change of the security's trading range. Developed by Marc Chaikin.

Usage

chaikinVolatility(HL, n = 10, maType, ...)

Arguments

HL
Object that is coercible to xts or matrix and contains High-Low prices.

n
Number of periods for moving average.

maType
A function or a string naming the function to be called.

...
Other arguments to be passed to the maType function.

Details

The Chaikin Volatility indicator defines volatility as an increase in the difference between the high and low.

Value

A object of the same class as HL or a vector (if try.xts fails) containing the Chaikin Volatility values.

Note

A rapid increase in Chaikin Volatility indicates an approaching bottom. A slow decrease in Chaikin Volatility indicates an approaching top.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.fmlabs.com/reference/ChaikinVolatility.htm
https://www.metastock.com/Customer/Resources/TAAZ/?p=120

See Also

See EMA, SMA, etc. for moving average options; and note Warning section. See TR for another volatility measure.

Examples

data(ttrc)
volatility <- chaikinVolatility(ttrc[,c("High","Low")])

[Package TTR version 0.23-2 Index]