ttrc {TTR}	R Documentation

Technical Trading Rule Composite data

Description

Historical Open, High, Low, Close, and Volume data for the periods January 2, 1985 to December 31, 2006. Randomly generated.

Format

The format is:

Date: 	 Class 'Date' 	 5480 5481 5482 5485 5486 ...

Open: 	 num 	 3.18 3.09 3.11 3.09 3.10 ...

High: 	 num 	 3.18 3.15 3.12 3.12 3.12 ...

Low: 	 num 	 3.08 3.09 3.08 3.07 3.08 ...

Close: 	 num 	 3.08 3.11 3.09 3.10 3.11 ...

Volume: 	 num 	 1870906 3099506 2274157 2086758 2166348 ...

Details

These data do not represent an actual security. They are provided so examples do not necessitate an internet connection.

Source

Randomly generated.

Examples

data(ttrc)
plot(tail(ttrc[,"Close"],100), type="l")

[Package TTR version 0.23-2 Index]