ultimateOscillator {TTR}	R Documentation

The Ultimate Oscillator

Description

The Ultimate Oscillator is a momentum oscillator designed to capture momentum across three different time frames.

Usage

ultimateOscillator(HLC, n = c(7, 14, 28), wts = c(4, 2, 1))

Arguments

HLC
Object that is coercible to xts or matrix and contains High-Low-Close prices.

n
A vector of the number of periods to use for each average calculation.

wts
The weights applied to each average.

Details

Created by Larry Williams in 1976.

Author(s)

Ivan Popivanov

References

The following site(s) were used to code/document this indicator:
http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:ultimate_oscillator

Examples

data(ttrc)
ult.osc <- ultimateOscillator(ttrc[,c("High","Low","Close")])

[Package TTR version 0.23-2 Index]