williamsAD {TTR}	R Documentation

Williams Accumulation / Distribution

Description

The Williams Accumulation / Distribution (AD) line is a measure of market momentum. Developed by Larry Williams.

Usage

williamsAD(HLC)

Arguments

HLC
Object that is coercible to xts or matrix and contains High-Low-Close prices.

Details

The Williams AD line differs from OBV and chaikinAD in that it doesn't take volume into account.

Value

A object of the same class as HLC or a vector (if try.xts fails) containing the accumulation / distribution values.

Note

The Accumulation/Distribution Line is interpreted by looking for a divergence in the direction of the indicator relative to price.

Author(s)

Joshua Ulrich

References

The following site(s) were used to code/document this indicator:
http://www.fmlabs.com/reference/WilliamsAD.htm
http://www.metastock.com/Customer/Resources/TAAZ/#125

See Also

See OBV, chaikinAD, and ATR.

Examples

data(ttrc)
ad <- williamsAD(ttrc[,c("High","Low","Close")])

[Package TTR version 0.23-2 Index]